sentry: address PR review feedback
