+++ GraceNotes/GraceNotes/Features/Journal/Services/ShareCardStyle.swift
@@ -1,16 +1,5 @@
 import SwiftUI
 
-private extension Color {
-    /// 24-bit `RRGGBB`. High bits are masked so `AARRGGBB`-style literals still resolve to the same sRGB triplet.
-    init(hex: UInt) {
-        let rgb = hex & 0xFFFFFF
-        let red = Double((rgb >> 16) & 0xFF) / 255
-        let green = Double((rgb >> 8) & 0xFF) / 255
-        let blue = Double(rgb & 0xFF) / 255
-        self.init(.sRGB, red: red, green: green, blue: blue)
-    }
-}
-
 /// Named presets for the exported share card bitmap (Figma Make: grace-notes / editorial / embellished).
 enum ShareCardStyle: String, CaseIterable, Identifiable, Sendable {
     case paperWarm
