+++ GraceNotes/GraceNotes/Features/Journal/Services/ShareCardSurface.swift
@@ -109,12 +109,3 @@ private func darkCompletionChipBackground(style: ShareCardStyle) -> some View {
         )
     }
 }
-
-private extension Color {
-    init(hex: UInt) {
-        let red = Double((hex >> 16) & 0xFF) / 255
-        let green = Double((hex >> 8) & 0xFF) / 255
-        let blue = Double(hex & 0xFF) / 255
-        self.init(red: red, green: green, blue: blue)
-    }
-}
